+++ Final/src/main/java/com/smhrd/controller/MainController.java
@@ -9,9 +9,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.smhrd.db.MemberMapper;
 import com.smhrd.model.MemberVO;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @Controller
 public class MainController {
@@ -32,8 +36,10 @@ public class MainController {
 	@PostMapping("/joinProcess")
 	public String join(MemberVO vo, HttpSession session) {
 		mapper.join(vo);
+		vo.setMemPw(null);
+		vo.setGender(null);
 		session.setAttribute("member", vo);
-		return "redirect:/";
+		return "redirect:/main";
 	}
 
 	@PostMapping("/login")
@@ -66,34 +72,37 @@ public class MainController {
 	@PostMapping("/update")
 	public String update(MemberVO vo, HttpSession session) {
 		MemberVO member = (MemberVO) session.getAttribute("member");
+//		System.out.println(member);
+//		System.out.println(vo);
 		
-		// 기존 이름, 성별, 생일, 가입일 정보 유지
-		vo.setMemId(member.getMemId());
-		if (vo.getName() == null) {
-			vo.setName(member.getName());
-		}
-		if (vo.getGender() == null) {
-			vo.setGender(member.getGender());
-		}
-		if (vo.getBirth() == null) {
-			vo.setBirth(member.getBirth());
-		}
-		if (vo.getCreatedAt() == null) {
-			vo.setCreatedAt(member.getCreatedAt());
-		}
-
 		mapper.update(vo);
-		session.setAttribute("member", vo);
+//		System.out.println(member);
 		return "redirect:/main";
 	}
+	
 
 	@PostMapping("/delete")
-	public String delete(HttpSession session) {
+	public String delete(MemberVO vo, HttpSession session) {
+	    Logger logger = LoggerFactory.getLogger(this.getClass());
+
 		MemberVO member = (MemberVO) session.getAttribute("member");
+		
+		System.out.println(member);
 		String memId = member.getMemId();
-		mapper.delete(memId);
-		session.invalidate();
-		return "redirect:/";
+		vo.setMemId(memId);
+
+		// 비밀번호 일치 확인
+	    int count = mapper.checkDelete(vo);
+
+	    logger.info("checkDelete 반환값: " + count); // 반환값 로그 출력
+
+	    if (count == 1) {
+        	mapper.delete(vo);
+        	session.invalidate();
+        	return "redirect:/";
+        }else {
+        	return "redirect:/mypage";
+        }
 	}
 
 }
