+++ Final/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -21,5 +21,5 @@
 		<beans:property name="suffix" value=".jsp" />
 	</beans:bean>
 
-	<context:component-scan base-package="com.smhrd.controller" />
+	<context:component-scan base-package="com.smhrd.controller, com.smhrd.restcontroller" />
 </beans:beans>
